Made some updates

+++ contracts/CapitalTechCrowdsale.sol
@@ -3,17 +3,14 @@ Capital Technologies & Research - Capital (CALL) & CapitalGAS (CALLG) - Crowdsal
 https://www.mycapitalco.in
 */
 pragma solidity 0.4.24;
-
 import './CALLGToken.sol';
 import './CALLToken.sol';
 import './TeamVault.sol';
 import './BountyVault.sol';
 import 'openzeppelin-solidity/contracts/crowdsale/distribution/utils/RefundVault.sol';
-
 contract FiatContract {
-  function USD(uint256 _id) public constant returns (uint256);
+  function USD(uint _id) public constant returns (uint256);
 }
-
 contract CapitalTechCrowdsale is Ownable {
   using SafeMath for uint256;
   ERC20 public token_call;
